Add DeletePrescription endpoint to ReservationController
A new HTTP DELETE endpoint, `DeletePrescription`, was added to the
`ReservationController` in the `Presentation.Controllers` namespace.
This method allows users with the "doctor" role to delete prescriptions
associated with appointments. It is secured with `[Authorize(Roles = "doctor")]`
and utilizes the `_serviceManager.AppointmentOrchestrator` service to
perform the deletion. The method returns a `NoContent` response upon
successful execution.